chore: move controls to separate folder
