+++ frontend/src/components/controls/ColorPicker.vue
