+++ frontend/src/components/controls/NumberInput.vue
