+++ frontend/src/components/controls/SliderInput.vue
